+++ cocos/web-view/web-view.ts
@@ -24,15 +24,15 @@
  */
 
 /**
- * @category component/webview
+ * @category component/web-view
  */
 
 import { ccclass, help, executeInEditMode, menu, tooltip, type, displayOrder, serializable, requireComponent } from 'cc.decorator';
 import { EDITOR } from 'internal:constants';
 import { UITransform } from '../core/components/ui-base';
 import { Component, EventHandler as ComponentEventHandler } from '../core/components';
-import { WebViewImplManager } from './webview-Impl-manager';
-import { EventType } from './webview-enums';
+import { WebViewImplManager } from './web-view-impl-manager';
+import { EventType } from './web-view-enums';
 import { legacyCC } from "../core/global-exports";
 
 /**
